refactor: change base button to n-button
