+++ frontend/src/components/common/AppHeader.vue
