+++ frontend/src/components/common/LoadingSpinner.vue
