+++ frontend/src/layout/DefaultLayout.vue
@@ -8,5 +8,5 @@
 </template>
 
 <script setup lang="ts">
-import AppHeader from '@/components/AppHeader.vue';
+import AppHeader from '@/components/common/AppHeader.vue';
 </script>
